+++ tests/testthat/tests.make_standata.R
@@ -231,17 +231,33 @@ test_that("make_standata handles multivariate models", {
   )
   expect_true(all(sdata_names %in% names(sdata)))
   expect_equal(sdata$con_theta_x, as.array(c(2, 1)))
+})
+
+test_that("make_standata handles the 'subset' addition argument correctly", {
+  dat1 <- data.frame(
+    y1 = rnorm(15), y2 = NA,
+    x1 = rnorm(15), x2 = NA, x3 = rnorm(15),
+    sub1 = 1, sub2 = 0
+  )
+  dat2 <- data.frame(
+    y1 = NA, y2 = rnorm(10),
+    x1 = NA, x2 = rnorm(10), x3 = NA,
+    sub1 = 0, sub2 = 1
+  )
+  dat <- rbind(dat1, dat2)
   
-  # test addition argument 'subset'
-  bform <- bf(y1 | subset(censi) ~ x + y2 + (x|2|g)) + 
-    (bf(y2 ~ s(y2) + (1|2|g)) + skew_normal())
+  bform <- 
+    bf(y1 | subset(sub1) ~ x1*x3 + sin(x1), family = gaussian()) +
+    bf(y2 | subset(sub2) ~ x2, family = gaussian()) +
+    set_rescor(FALSE)
+  
   sdata <- make_standata(bform, dat)
-  nsub <- sum(dat$censi)
-  expect_equal(sdata$N_y1, nsub)
-  expect_equal(sdata$N_y2, nrow(dat))
-  expect_equal(length(sdata$Y_y1), nsub)
-  expect_equal(nrow(sdata$X_y1), nsub)
-  expect_equal(length(sdata$Z_1_y1_2), nsub)
+  nsub1 <- sum(dat$sub1)
+  nsub2 <- sum(dat$sub2)
+  expect_equal(sdata$N_y1, nsub1)
+  expect_equal(sdata$N_y2, nsub2)
+  expect_equal(length(sdata$Y_y1), nsub1)
+  expect_equal(nrow(sdata$X_y2), nsub2)
 })
 
 test_that("make_standata returns correct data for ARMA terms", {